Fix: Prevent tuning of parameters on whole dataset
As `auto_arima` is run on the whole dataset, there is leakage of
algorithm parameters and the accuracy score for the test set might be
inflated as the test set was used to chose the ARIMA parameters. It's
for this same reason that Kaggle has a private leaderboard etc. In order
to prevent this I've simply split the data between train and test before
tuning the parameters.